Cleanup configuration tree
Sensors mapping is now done in configuration files instead of
mapping files. This is mainly due to the fact that the mapping of
the sensors is decided at rythmus implentation, and should not be
modified during the mapping.
If two sensors are inverted, the saner thing to do is to invert
both data cables, instead of fiddling with the mapping.