Make sensor cycles threshold adptative
Before starting, the sensors will now compute the median of N samples (N
being defined as `calibratedCyclesSamplesLength` in .apprc) to determine
the ideal cyclesThreshold. The sensor led will quickly blink during this
calibration phase.

This will allow Rythmus to run in various conditions without the need to
update a hardcoded value.

Note that if the conditions change during a run, the install will need
to be rebooted to update the calibrated threshold.